move plugins/BaseMVAValueMapProducer.h to interface/
