+++ PhysicsTools/PatAlgos/BuildFile.xml
@@ -12,8 +12,9 @@
 <use name="FWCore/Framework"/>
 <use name="FWCore/ParameterSet"/>
 <use name="FWCore/Utilities"/>
-<use name="PhysicsTools/PatUtils"/>
 <use name="PhysicsTools/IsolationAlgos"/>
+<use name="PhysicsTools/PatUtils"/>
+<use name="PhysicsTools/TensorFlow"/>
 <use name="clhep"/>
 <export>
   <lib name="1"/>
